etl/requirements.txt
openhexa-sdk
psycopg2-binary
pandas
sqlalchemy